move check analyze from analyzer to planner
