+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/BinaryArithmetic.java
@@ -17,25 +17,8 @@
 
 package org.apache.doris.nereids.trees.expressions;
 
-import org.apache.doris.nereids.trees.expressions.shape.LeafExpression;
-
-import java.util.List;
-
-/**
- * Abstract class for all slot in expression.
- */
-public abstract class Slot extends NamedExpression implements LeafExpression {
-
-    @Override
-    public Slot toSlot() {
-        return this;
-    }
-
-    public Slot withNullable(boolean newNullable) {
-        throw new RuntimeException("Do not implement");
-    }
-
-    public Slot withQualifier(List<String> qualifiers) {
-        throw new RuntimeException("Do not implement");
+public abstract class BinaryArithmetic extends BinaryOperator {
+    public BinaryArithmetic(Expression left, Expression right, String symbol) {
+        super(left, right, symbol);
     }
 }
