+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/CompoundPredicate.java
@@ -22,29 +22,21 @@ import org.apache.doris.nereids.trees.expressions.visitor.ExpressionVisitor;
 import org.apache.doris.nereids.types.BooleanType;
 import org.apache.doris.nereids.types.DataType;
 
-import java.util.Objects;
-
 /**
  * Compound predicate expression.
  * Such as &&,||,AND,OR.
  */
-public abstract class CompoundPredicate extends Expression implements BinaryExpression {
-    protected final String symbol;
+public abstract class CompoundPredicate extends BinaryOperator {
 
     /**
      * Desc: Constructor for CompoundPredicate.
      *
      * @param left  left child of comparison predicate
      * @param right right child of comparison predicate
+     * @param symbol symbol used in sql
      */
     public CompoundPredicate(Expression left, Expression right, String symbol) {
-        super(left, right);
-        this.symbol = symbol;
-    }
-
-    @Override
-    public String toSql() {
-        return "(" + left().toSql() + " " + symbol + " " + right().toSql() + ")";
+        super(left, right, symbol);
     }
 
     @Override
@@ -60,29 +52,6 @@ public abstract class CompoundPredicate extends Expression implements BinaryExpr
     @Override
     public <R, C> R accept(ExpressionVisitor<R, C> visitor, C context) {
         return visitor.visitCompoundPredicate(this, context);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        CompoundPredicate other = (CompoundPredicate) o;
-        return Objects.equals(left(), other.left())
-                && Objects.equals(right(), other.right());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(symbol, left(), right());
-    }
-
-    @Override
-    public String toString() {
-        return "(" + left().toString() + " " + symbol + " " + right().toString() + ")";
     }
 
     /**
